refactor: split handlers/request.ts into three files by passage role
`handlers/request.ts` had grown to 1656 lines containing 10+ verbs
and helpers spanning four passage roles. Split it into:

- request.ts             (605 lines) — create + fast-track + their
                                       shared parsers (parseExecutorsList,
                                       parseWithList)
- requestReads.ts        (582 lines) — list/pending/show + text-mode
                                       formatters (formatRequestText,
                                       printSummary, formatReviewMarkers,
                                       computeReviewMarkerWidth,
                                       describeFilters)
- requestLifecycle.ts    (533 lines) — approve/deny/execute/complete/fail
                                       + lifecycle helpers
                                       (rejectIfNonMember, emitSliceClose,
                                       dashedValueHint, resolveReason)

Each file is now under 700 lines and groups handlers by passage role
(create / read / lifecycle / fast-track). No behavior change — every
handler's signature, output, and lifecycle remain byte-identical.
Full test suite passes unchanged (1677/1677).

Import path updates:
- src/interface/gate/index.ts — three separate imports
- src/interface/gate/handlers/board.ts — formatReviewMarkers /
  computeReviewMarkerWidth from requestReads.js
- handlers/issues.ts — parseExecutorsList still in request.ts
  (unchanged)
- The test-backward-compat re-export in index.ts redirects to the
  new requestReads.js path so tests/interface/reviewMarkers.test.ts
  still imports from `../../src/interface/gate/index.js`.
